1.19.4 port
BUILD: Updated Minecraft version from 1.19.3 to 1.19.4
BUILD: Updated all dependencies to 1.19.4 (Fabric, Fabric API, etc)
BUILD: Updated all mods which would be used in dev env to 1.19.4(YACL, Mod Menu)
REFACTOR: Updated code to 1.19.4

Oh how I've missed the QoL changes for 1.19.4 for both the creative mode inventory rework and Mod Menu changes

+++ src/main/java/net/superkat/explosiveenhancement/mixin/ExplosionMixin.java
@@ -27,7 +27,7 @@ public abstract class ExplosionMixin {
 		if(ExplosiveConfig.debugLogs) {
 			LOGGER.info("affectWorld has been called!");
 		}
-		BlockPos pos = new BlockPos(x, y, z);
+		BlockPos pos = new BlockPos((int) x, (int) y, (int) z);
 		if(world.getFluidState(pos).isIn(FluidTags.WATER) && ExplosiveConfig.underwaterExplosions) {
 			//If underwater
 			isUnderWater = true;
